fix ctl-reconfigure issue (can't find sqitch.plan)
ERROR:

    chef-server:       [execute] Possible precedence issue with control flow operator at /opt/opscode/embedded/lib/perl5/site_perl/5.34.0/App/Sqitch/Plan.pm line 578.
    chef-server:                 Plan file /opt/opscode/embedded/service/opscode-erchef/schema/baseline/sqitch.plan does not exist

FILES EXIST AT:
/opt/opscode/embedded/service/opscode-erchef/baseline/ and /opt/opscode/embedded/service/opscode-erchef/ vs /opt/opscode/embedded/service/opscode-erchef/schema/baseline/

FIX:
in the dev vm, i created the directory it was looking for, and moved the sqitch.plan file at /opt/opscode/embedded/service/opscode-erchef/baseline there, and did the reconfigure.  that worked, and brought us to the next error.

remove the word 'schema' from the directory paths.

Signed-off-by: Lincoln Baker <lbaker@chef.io>